--- app/ns-ui-widgets-category/progress/basics/article.md
@@ -1,5 +1,0 @@
-The example shows how to create basic progress component as well as, how to change its value property via code-behind. 
-
-<snippet id='progress-create-xml'/>
-<snippet id='progress-value-change-event'/>
-<snippet id='progress-value-change-event-ts'/>
