--- app/ns-ui-widgets-category/progress/code-behind/article.md
@@ -1,6 +1,0 @@
-In the following example is shown, how to create Progress via Code-behind
-
-<snippet id='creating-progress-xml'/>
-
-<snippet id='creating-progress-code'/>
-<snippet id='creating-progress-code-ts'/>
