+++ app/ns-ui-widgets-category/progress/end.md
@@ -1,8 +1,20 @@
 
-**API Reference for** [Progress Class](http://docs.nativescript.org/api-reference/modules/_ui_progress_.html)
+## Properties
 
-**Native Component**
+| Name     | Type    | Description    |
+|----------|---------|----------------|
+| `value`   | `number` | Gets or sets a progress current value. |
+| `maxValue`   | `number` | Gets or sets a progress max value. |
 
-| Android                | iOS      |
-|:-----------------------|:---------|
-| [android.widget.ProgressBar](http://developer.android.com/reference/android/widget/ProgressBar.html) (indeterminate = false) | [UIProgressView](https://developer.apple.com/library/ios/documentation/UIKit/Reference/UIProgressView_Class/) |
+## API References
+
+| Name     | Type    | 
+|----------|---------|
+| [tns-core-modules/ui/progress](http://docs.nativescript.org/api-reference/modules/_ui_progress_.html) | `Module` | 
+| [Progress](https://docs.nativescript.org/api-reference/classes/_ui_progress_.progress) | `Class` | 
+
+## Native Component
+
+| Android               | iOS      |
+|:----------------------|:---------|
+| [android.widget.ProgressBar](http://developer.android.com/reference/android/widget/ProgressBar.html) | [UIProgressView](https://developer.apple.com/library/ios/documentation/UIKit/Reference/UIProgressView_Class/) |
