+++ app/ns-ui-widgets-category/progress/metadata.md
@@ -4,3 +4,4 @@ description: The Progress class provides an abstraction over the iOS's UIProgres
 position: 340
 slug: progress
 ---
+example-order: usage, styling, tips-and-tricks
