+++ app/ns-ui-widgets-category/progress/overview.md
@@ -1,5 +1,2 @@
-The Progress widget is a visual bar indicator of a progress in a operation. 
+The `Progress` widget is a visual bar indicator of a progress in a operation. 
 It shows a bar representing the current progress of the operation.
-
-<snippet id='progress-require'/>
-<snippet id='progress-import'/>
