+++ app/ns-ui-widgets-category/progress/progress-page.js
@@ -1,12 +1,14 @@
 const ListViewLinksModel = require("../../links-view-model");
 const link = require("../../link");
 const navigationLinks = [
-    new link("Basics", "ns-ui-widgets-category/progress/basics/basics-page"),
+    new link("Usage", "ns-ui-widgets-category/progress/usage/usage-page"),
+    new link("Tips and tricks", "ns-ui-widgets-category/progress/tips-and-tricks/tips-and-tricks-page"),
     new link("Code Behind", "ns-ui-widgets-category/progress/code-behind/code-behind-page"),
-    new link("Style", "ns-ui-widgets-category/progress/style/style-page")
+    new link("Style", "ns-ui-widgets-category/progress/styling/styling-page")
 ];
 const navigationLinksTsc = [
-    new link("Basics", "ns-ui-widgets-category/progress/basics/basics-ts-page"),
+    new link("Usage", "ns-ui-widgets-category/progress/usage/usage-ts-page"),
+    new link("Tips and tricks", "ns-ui-widgets-category/progress/tips-and-tricks/tips-and-tricks-ts-page"),
     new link("Code Behind", "ns-ui-widgets-category/progress/code-behind/code-behind-ts-page")
 ];
 function onNavigatingTo(args) {
