--- app/ns-ui-widgets-category/progress/style/article.md
@@ -1,4 +1,0 @@
-The example shows how to style progress component.
-
-<snippet id='progress-style-xml'/>
-<snippet id='progress-style-css'/>
