--- app/ns-ui-widgets-category/progress/style/style-page.css
@@ -1,21 +1,0 @@
-button, label, stack-layout {
-    horizontal-align: center;    
-}
-
-.icon {
-   font-family: 'icomoon';
-}
-
-GridLayout {
-    padding: 0;
-    margin-left: 0;
-    margin-top: 10;
-    margin-bottom: 10;
-}
-/* >> progress-style-css */
-Progress{
-    color: cyan;
-    background-color: green;
-    
-}
-/* << progress-style-css */
