--- app/ns-ui-widgets-category/progress/style/style-page.xml
@@ -1,12 +1,0 @@
-<Page xmlns="http://www.nativescript.org/tns.xsd" navigatingTo="onNavigatingTo" loaded="onPageLoaded">
-    <Page.actionBar>
-		<ActionBar title="Getting Started"/>
-	</Page.actionBar>
-    <GridLayout rows="*">
-        <StackLayout verticalAlign="center" row="0" height="100">
-            <!-- >> progress-style-xml -->
-            <Progress width="100%" value="{{ progressValue }}"  maxValue="{{ progressMaxValue }}" loaded="onProgressLoaded" class="p-20"/>
-            <!-- << progress-style-xml -->
-        </StackLayout>
-    </GridLayout>
-</Page>
