+++ app/ns-ui-widgets-category/progress/styling/styling-page.css
@@ -1,0 +1,7 @@
+/* >> progress-style-css */
+Progress{
+    color: cyan;
+    background-color: green;
+    
+}
+/* << progress-style-css */
