+++ app/ns-ui-widgets-category/progress/styling/styling-page.xml
@@ -1,0 +1,18 @@
+<Page xmlns="http://www.nativescript.org/tns.xsd">
+    <Page.actionBar>
+		<ActionBar title="Styling"/>
+	</Page.actionBar>
+    <GridLayout rows="*">
+        <GridLayout rows="* *" height="150" class="p-20">
+            <!-- >> progress-style-xml -->
+            <!-- 
+                Using backgroundColor (CSS: background-color) & color to change the Progress style 
+                Note; backgroundColor will work only on iOS; onAndroid the background will be the color with applied opacity
+            -->
+            <Progress value="50" maxValue="100" backgroundColor="red" color="green" row="0"/>
+            <!-- Using the nativescript-theme-core CSS class to change the Progress style -->
+            <Progress value="25" maxValue="100" class="progress" row="1"/>
+            <!-- << progress-style-xml -->
+        </GridLayout>
+    </GridLayout>
+</Page>
