+++ app/ns-ui-widgets-category/progress/tips-and-tricks/article.md
@@ -1,0 +1,8 @@
+<snippet id='progress-create-xml'/>
+<snippet id='progress-value-change-event'/>
+<snippet id='progress-value-change-event-ts'/>
+
+### Creating Progress component via Code-Behind
+
+<snippet id='progress-style-xml'/>
+<snippet id='progress-style-css'/>
