+++ app/ns-ui-widgets-category/progress/tips-and-tricks/tips-and-tricks-page.xml
@@ -1,0 +1,12 @@
+<Page xmlns="http://www.nativescript.org/tns.xsd" navigatingTo="onNavigatingTo" loaded="onPageLoaded">
+    <Page.actionBar>
+		<ActionBar title="Getting Started"/>
+	</Page.actionBar>
+    <GridLayout rows="*">
+        <!-- >> progress-create-xml -->
+        <StackLayout verticalAlign="center" height="100">
+            <Progress width="100%" value="{{ progressValue }}"  maxValue="{{ progressMaxValue }}" loaded="onProgressLoaded" />
+        </StackLayout>
+        <!-- << progress-create-xml -->
+    </GridLayout>
+</Page>
