+++ app/ns-ui-widgets-category/progress/usage/article.md
@@ -1,0 +1,3 @@
+<snippet id='progress-create-xml'/>
+<snippet id='progress-value-change-event'/>
+<snippet id='progress-value-change-event-ts'/>
