+++ app/ns-ui-widgets-category/progress/usage/usage-page.xml
@@ -1,11 +1,11 @@
 <Page xmlns="http://www.nativescript.org/tns.xsd" navigatingTo="onNavigatingTo" loaded="onPageLoaded">
     <Page.actionBar>
-		<ActionBar title="Getting Started"/>
+		<ActionBar title="Usage"/>
 	</Page.actionBar>
     <GridLayout rows="*">
         <!-- >> progress-create-xml -->
-        <StackLayout verticalAlign="center" row="0" height="100">
-            <Progress width="100%" value="{{ progressValue }}"  maxValue="{{ progressMaxValue }}" loaded="onProgressLoaded" class="p-20"/>
+        <StackLayout verticalAlign="center" height="100" class="p-20">
+            <Progress value="25" maxValue="100"  loaded="onProgressLoaded"/>
         </StackLayout>
         <!-- << progress-create-xml -->
     </GridLayout>
