+++ app/ns-ui-widgets-category/progress/usage/usage-ts-page.xml
@@ -1,0 +1,10 @@
+<Page xmlns="http://www.nativescript.org/tns.xsd" navigatingTo="onNavigatingTo" loaded="onPageLoaded">
+    <Page.actionBar>
+		<ActionBar title="Usage"/>
+	</Page.actionBar>
+    <GridLayout rows="*">
+        <StackLayout verticalAlign="center" height="100" class="p-20">
+            <Progress value="25" maxValue="100"  loaded="onProgressLoaded"/>
+        </StackLayout>
+    </GridLayout>
+</Page>
