+++ app/ns-ui-widgets-category/scroll-view/end.md
@@ -1,2 +1,26 @@
+## Properties
 
-**API Reference for the** [ScrollView Class](http://docs.nativescript.org/api-reference/modules/_ui_scroll_view_.html)
+| Name     | Type    | Description    |
+|----------|---------|----------------|
+| `orientation`   | [`Orientation`](https://docs.nativescript.org/api-reference/modules/_ui_scroll_view_#orientation) | Gets or sets direction in which the content can be scrolled. Supported values are _"vertical"_ (default) and _"horizontal"_. |
+
+## Events
+
+| Name     | Description    |
+|----------|----------------|
+| `scroll`    | Emitted when the component is scrolled. |
+
+## API References
+
+| Name     | Type    | 
+|----------|---------|
+| [tns-core-modules/ui/scroll-view](http://docs.nativescript.org/api-reference/modules/_ui_scroll_view_.html) | `Module` | 
+| [ScrollView](https://docs.nativescript.org/api-reference/classes/_ui_scroll_view_.scrollview) | `Class` | 
+| [orientation](https://docs.nativescript.org/api-reference/classes/_ui_scroll_view_.scrollview#orientation) |`Property`|
+| [ScrollEventData](https://docs.nativescript.org/api-reference/interfaces/_ui_scroll_view_.scrolleventdata) | `Interface` |
+
+## Native Component
+
+| Android               | iOS      |
+|:----------------------|:---------|
+| _NativeScript specific implementation_ | _NativeScript specific implementation_ |
