--- app/ns-ui-widgets-category/scroll-view/events/article.md
@@ -1,5 +1,0 @@
-## Scroll events
-
-<snippet id='scroll-event-xml'/>
-<snippet id='scroll-view-event-code'/>
-<snippet id='scroll-view-event-code-ts'/>
