--- app/ns-ui-widgets-category/scroll-view/events/events-page.xml
@@ -1,32 +1,0 @@
-<Page xmlns="http://www.nativescript.org/tns.xsd" navigatingTo="onNavigatingTo">
-    <Page.actionBar>
-		<ActionBar title="Events"/>
-	</Page.actionBar>
-    <StackLayout>
-        <!-- >> scroll-event-xml -->
-        <StackLayout borderRadius="2" backgroundColor="#0099CC" class="m-b-5" height="50">
-            <Label text="{{ status }}" class="h2 text-center" color="white"></Label>
-        </StackLayout>
-        <ScrollView scroll="onScroll">
-            <GridLayout class="m-15" rows="auto auto auto auto auto auto auto auto auto auto auto auto auto auto auto auto">
-                <Label class="h3 m-5" height="50" row="0" text="Title 1"/>
-                <Label class="h3 m-5" height="50" row="1" text="Title 2"/>
-                <Label class="h3 m-5" height="50" row="2" text="Title 3"/>
-                <Label class="h3 m-5" height="50" row="3" text="Title 4"/>
-                <Label class="h3 m-5" height="50" row="4" text="Title 5"/>
-                <Label class="h3 m-5" height="50" row="5" text="Title 6"/>
-                <Label class="h3 m-5" height="50" row="6" text="Title 7"/>
-                <Label class="h3 m-5" height="50" row="7" text="Title 8"/>
-                <Label class="h3 m-5" height="50" row="8" text="Title 9"/>
-                <Label class="h3 m-5" height="50" row="9" text="Title 10"/>
-                <Label class="h3 m-5" height="50" row="10" text="Title 10"/>
-                <Label class="h3 m-5" height="50" row="11" text="Title 11"/>
-                <Label class="h3 m-5" height="50" row="12" text="Title 12"/>
-                <Label class="h3 m-5" height="50" row="13" text="Title 13"/>
-                <Label class="h3 m-5" height="50" row="14" text="Title 14"/>
-                <Label class="h3 m-5" height="50" row="15" text="Title 15"/>
-            </GridLayout>
-        </ScrollView>
-        <!-- << scroll-event-xml -->
-    </StackLayout>
-</Page>
