--- app/ns-ui-widgets-category/scroll-view/horizontal/article.md
@@ -1,3 +1,0 @@
-Set ScrollView with `horizontal` orientation.
-
-<snippet id='scroll-view-horizontal-xml'/>
