--- app/ns-ui-widgets-category/scroll-view/horizontal/horizontal-page.js
@@ -1,5 +1,0 @@
-function onPageLoaded(args) {
-   console.log("Horizontal ScrollView Page loaded");
-}
-
-exports.onPageLoaded = onPageLoaded;
