--- app/ns-ui-widgets-category/scroll-view/horizontal/horizontal-page.xml
@@ -1,27 +1,0 @@
-<Page xmlns="http://www.nativescript.org/tns.xsd" loaded="onPageLoaded">
-    <Page.actionBar>
-      <ActionBar title="Horizontal"/>
-    </Page.actionBar>
-    <!-- >> scroll-view-horizontal-xml -->
-    <ScrollView orientation="horizontal">
-        <GridLayout class="m-15" columns="auto auto auto auto auto auto auto auto auto auto auto auto auto auto auto auto">
-            <Label width="50" height="50" class="h3 m-15" col="0" textWrap="true" text="Title 1"/>
-            <Label width="50" height="50" class="h3 m-15" col="1" textWrap="true" text="Title 2"/>
-            <Label width="50" height="50" class="h3 m-15" col="2" textWrap="true" text="Title 3"/>
-            <Label width="50" height="50" class="h3 m-15" col="3" textWrap="true" text="Title 4"/>
-            <Label width="50" height="50" class="h3 m-15" col="4" textWrap="true" text="Title 5"/>
-            <Label width="50" height="50" class="h3 m-15" col="5" textWrap="true" text="Title 6"/>
-            <Label width="50" height="50" class="h3 m-15" col="6" textWrap="true" text="Title 7"/>
-            <Label width="50" height="50" class="h3 m-15" col="7" textWrap="true" text="Title 8"/>
-            <Label width="50" height="50" class="h3 m-15" col="8" textWrap="true" text="Title 9"/>
-            <Label width="50" height="50" class="h3 m-15" col="9" textWrap="true" text="Title 10"/>
-            <Label width="50" height="50" class="h3 m-15" col="10" textWrap="true" text="Title 10"/>
-            <Label width="50" height="50" class="h3 m-15" col="11" textWrap="true" text="Title 11"/>
-            <Label width="50" height="50" class="h3 m-15" col="12" textWrap="true" text="Title 12"/>
-            <Label width="50" height="50" class="h3 m-15" col="13" textWrap="true" text="Title 13"/>
-            <Label width="50" height="50" class="h3 m-15" col="14" textWrap="true" text="Title 14"/>
-            <Label width="50" height="50" class="h3 m-15" col="15" textWrap="true" text="Title 15"/>
-        </GridLayout>
-    </ScrollView>
-    <!-- << scroll-view-horizontal-xml -->
-</Page>
