--- app/ns-ui-widgets-category/scroll-view/is-scroll-enabled/article.md
@@ -1,5 +1,0 @@
-The `isScrollEnabled` property allows you to enable/disable the  scroll view of the ScrollView component. The example shows how to bind the property and how to toggle its value.
-
-<snippet id='scroll-view-is-enabled-xml'/>
-<snippet id='scroll-view-is-enabled-js'/>
-<snippet id='scroll-view-is-enabled-ts'/>
