+++ app/ns-ui-widgets-category/scroll-view/metadata.md
@@ -4,3 +4,4 @@ description: ScrollView class allows displaying scrollable area, which has conte
 position: 360
 slug: scroll-view
 ---
+example-order: usage, tips-and-tricks
