+++ app/ns-ui-widgets-category/scroll-view/scroll-view-page.js
@@ -1,14 +1,11 @@
 const ListViewLinksModel = require("../../links-view-model");
 const link = require("../../link");
 const navigationLinks = [
-    new link("Vertical", "ns-ui-widgets-category/scroll-view/vertical/vertical-page"),
-    new link("Horizontal", "ns-ui-widgets-category/scroll-view/horizontal/horizontal-page"),
-    new link("Events", "ns-ui-widgets-category/scroll-view/events/events-page"),
-    new link("Enable/Disable scroll", "ns-ui-widgets-category/scroll-view/is-scroll-enabled/is-scroll-enabled-page")
+    new link("Usage", "ns-ui-widgets-category/scroll-view/usage/usage-page"),
+    new link("Tips and tricks", "ns-ui-widgets-category/scroll-view/tips-and-tricks/tips-and-tricks-page")
 ];
 const navigationLinksTsc = [
-    new link("Events", "ns-ui-widgets-category/scroll-view/events/events-page"),
-    new link("Enable/Disable scroll", "ns-ui-widgets-category/scroll-view/is-scroll-enabled/is-scroll-enabled-ts-page")
+    new link("Tips and tricks", "ns-ui-widgets-category/scroll-view/tips-and-tricks/tips-and-tricks-ts-page")
 ];
 function onNavigatingTo(args) {
     const page = args.object;
