+++ app/ns-ui-widgets-category/scroll-view/tips-and-tricks/tips-and-tricks-page.js
@@ -1,0 +1,11 @@
+// >> scroll-view-event-code
+
+function onScroll(args) {
+    const scrollView = args.object;
+
+    console.log(`scrollX: ${args.scrollX}`);
+    console.log(`scrollY: ${args.scrollY}`);
+}
+
+exports.onScroll = onScroll;
+// << scroll-view-event-code
