+++ app/ns-ui-widgets-category/scroll-view/tips-and-tricks/tips-and-tricks-page.xml
@@ -1,0 +1,23 @@
+<Page xmlns="http://www.nativescript.org/tns.xsd">
+    <Page.actionBar>
+		<ActionBar title="Tips and Tricks"/>
+	</Page.actionBar>
+    <GridLayout>
+        <!-- >> scroll-event-xml -->
+        <ScrollView scroll="onScroll">
+            <GridLayout rows="200 200 200 200 200 200 200 200 200 200">
+                <Label row="0" text="Some row content goes here..."/>
+                <Label row="1" text="Some row content goes here..."/>
+                <Label row="2" text="Some row content goes here..."/>
+                <Label row="3" text="Some row content goes here..."/>
+                <Label row="4" text="Some row content goes here..."/>
+                <Label row="5" text="Some row content goes here..."/>
+                <Label row="6" text="Some row content goes here..."/>
+                <Label row="7" text="Some row content goes here..."/>
+                <Label row="8" text="Some row content goes here..."/>
+                <Label row="9" text="Some row content goes here..."/>
+            </GridLayout>
+        </ScrollView>
+        <!-- << scroll-event-xml -->
+    </GridLayout>
+</Page>
