+++ app/ns-ui-widgets-category/scroll-view/usage/article.md
@@ -1,0 +1,1 @@
+<snippet id='scroll-view-verical-xml'/>
