--- app/ns-ui-widgets-category/scroll-view/vertical/article.md
@@ -1,3 +1,0 @@
-Set ScrollView with `vertical` orientation.
-
-<snippet id='scroll-view-verical-xml'/>
