--- app/ns-ui-widgets-category/scroll-view/vertical/vertical-page.js
@@ -1,4 +1,0 @@
-function onPageLoaded(args) {
-    console.log("Vertical ScrollView Page loaded");
-}
-exports.onPageLoaded = onPageLoaded;
